fix server action
